restructuring the app routing
