+++ routes/debate.js
@@ -3,24 +3,6 @@ const router = express.Router();
 const path = require("path");
 const listing = require("../helpers/listings.js");
 
-// List of all debates
-router.get("/", (req, res) => {
-  try {
-    res.render("list", {
-      pageTitle: "Debates: All",
-      pageSubtitle: "List of all debates",
-      pageType: "home",
-      debateList: listing.getDebateListings("dummyData"),
-      user: {
-        username: req.cookies.s3o_username
-      }
-    });
-  } catch (err) {
-    console.error(err);
-    res.status(404).send("Sorry can't find that!");
-  }
-});
-
 // List of all debates of a specific type
 router.get("/:debateType", (req, res) => {
   try {
