--- views/index.hbs
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-{{#> head title='Debate demo home' subtitle='Debate demo home' type='home' }}
-{{/head}}
-
-<body>
-	<a href="/debate">Go to debates home/dashboard</a>
-	<footer>
-	</footer>
-</body>
-
-</html>
