+++ views/partials/user/header.hbs
@@ -1,6 +1,6 @@
 <header>
 	<div class="float-left">
-		<a href="/debate">Home</a>
+		<a href="/">Home</a>
 	</div>
 	<div class="float-right">
 		s3o as: {{ username }}
